feat(kvstore): Bulk Import (#927)
* refactor(undocumented): make abstraction more flexible

* refactor(secretstoreentry/create): move error type into shared location

* feat(kvstoreentry/create): implement buffered stdin read to undocumented ndjson endpoint

* refactor(kvstoreentry/create): remove debug info

* refactor(kvstoreentry/create): move stdin processing to separate function

* feat(kvstoreentry/create): implement streaming of json data via file

* feat(kvstoreentry/create): implement streaming of json data via directory structure

* fix(lint): gosec ignore

* fix(kvstore/create): base64 encode file contents and change key format

* temp(kvstoreentry): attempt to implement spinner into ProcessDir

The reason for the temp commit is that this work highlighted possible
performance concerns and so we'll want to rewrite this to avoid using
the batch API endpoint.

* refactor(kvstore/create): switch to concurrent calls for each key insertion

* refactor(kvstoreentry/create): remove unused args

* fix(kvstoreentry/create): add one second sleep to allow final goroutine to increment

* fix(lint): gosec ignore file read via variable

* docs(metadata): add examples for kvstoreentry create

* fix(kvstoreentry/list): paginate the results

* bump 8.0.3

* refactor(kvstoreentry/create): clean-up output messaging

* fix(metadata): incorrect link to KV Store APIs

* fix(lint): annotate methods

* build(deps): bump go-fastly to 8.1.0

* feat(kvstoreentry/create): switch to go-fastly implementation of batch endpoint

* fix(kvstoreentry/describe): render binary output safely

* fix(undocumented): Call() error return should include response data

* fix(undocumented): Call() to accept args as struct

* fix(kvstoreentry/create): ensure file is closed

* fix(kvstoreentry/create): inverse conditional to reduce indentation

* refactor(kvstoreentry/create): avoid explicit instantiation

* refactor(kvstoreentry/create): replace sleep with chan synchronisation

* refactor(kvstoreentry/create): inline variable

* fix(kvstoreentry/create): data race when error encountered

* fix(kvstoreentry/create): data race around mutating processing errors

* fix(kvstoreentry/create): retry API request if connection is lost

* fix(kvstoreentry/create): ignore hidden files

* feat(kvstoreentry/create): allow hidden files if flag is set

* fix(kvstoreentry/create): use latest go-fastly so we can pass io.Reader for buffering large files

* fix(kvstoreentry/create): add missing hidden build files

* test(kvstore): write tests for KV Store

* test(kvstoreentry): write tests for KV Store entries

* fix(kvstoreentry/create): ignore symlinks and correctly calculate files minus hidden files

* docs(kvstoreentry/create): document mutext protection

* refactor(undocumented): remove old func signature

* refactor(kvstoreentry/create): replace channel with atomic increment/load

* build: bump go-fastly to 8.3.0

* fix(kvstoreentry/create): use buffered io for key creation

* fix(compute/build): don't instrument build errors as they're not actionable

* test(kvstoreentry/create): add bulk import test cases

* style: make flag annoations consistent

* fix(kvstoreentry): use filepath package to handle path construction

* build: support callvis

* fix(kvstoreentry): correct the Windows logic for detecting hidden files

* fix(Makefile): avoid callvis .gv file generation and install errors

* fix(kvstoreentry): only check for dot in filename

Was unable to get checking the attributes of a file to work on Windows
platform in the CI environment.

* fix(kvstoreentry): implement Windows check for hidden file attribute

* Revert "fix(kvstoreentry): implement Windows check for hidden file attribute"

This reverts commit 24422570c89750988321949bfb8815bb62417083.

* docs(kvstoreentry/create): annotate exported ProcessErr type

+++ pkg/commands/kvstoreentry/delete.go
@@ -3,16 +3,20 @@ package kvstoreentry
 import (
 	"io"
 
+	"github.com/fastly/go-fastly/v8/fastly"
+
 	"github.com/fastly/cli/pkg/cmd"
+	fsterr "github.com/fastly/cli/pkg/errors"
 	"github.com/fastly/cli/pkg/global"
 	"github.com/fastly/cli/pkg/manifest"
 	"github.com/fastly/cli/pkg/text"
-	"github.com/fastly/go-fastly/v8/fastly"
 )
 
 // DeleteCommand calls the Fastly API to delete an kv store.
 type DeleteCommand struct {
 	cmd.Base
+	cmd.JSONOutput
+
 	manifest manifest.Data
 	Input    fastly.DeleteKVStoreKeyInput
 }
@@ -26,19 +30,43 @@ func NewDeleteCommand(parent cmd.Registerer, g *global.Data, m manifest.Data) *D
 		manifest: m,
 	}
 	c.CmdClause = parent.Command("delete", "Delete a key")
+
+	// Required.
+	c.CmdClause.Flag("key", "Key name").Short('k').Required().StringVar(&c.Input.Key)
 	c.CmdClause.Flag("store-id", "Store ID").Short('s').Required().StringVar(&c.Input.ID)
-	c.CmdClause.Flag("key-name", "Key name").Short('k').Required().StringVar(&c.Input.Key)
+
+	// Optional.
+	c.RegisterFlagBool(c.JSONFlag()) // --json
+
 	return &c
 }
 
 // Exec invokes the application logic for the command.
 func (c *DeleteCommand) Exec(_ io.Reader, out io.Writer) error {
+	if c.Globals.Verbose() && c.JSONOutput.Enabled {
+		return fsterr.ErrInvalidVerboseJSONCombo
+	}
+
 	err := c.Globals.APIClient.DeleteKVStoreKey(&c.Input)
 	if err != nil {
 		c.Globals.ErrLog.Add(err)
 		return err
 	}
 
-	text.Success(out, "Deleted key %s from store ID %s", c.Input.Key, c.Input.ID)
+	if c.JSONOutput.Enabled {
+		o := struct {
+			Key     string `json:"key"`
+			ID      string `json:"id"`
+			Deleted bool   `json:"deleted"`
+		}{
+			c.Input.Key,
+			c.Input.ID,
+			true,
+		}
+		_, err := c.WriteJSON(out, o)
+		return err
+	}
+
+	text.Success(out, "Deleted key %s from KV Store %s", c.Input.Key, c.Input.ID)
 	return nil
 }
